+++ tests/test_units/test_mapper_str.py
@@ -9,10 +9,10 @@ class TestMapperStr(unittest.TestCase):
         m.connect('entry', '/entries/{id}', controller='entry', action='show')
 
         expected = """\
-Route name Methods Path
+Route name Methods Path                   Controller action
                    /{controller}/{action}
-entries            /entries
-entry              /entries/{id}"""
-        
+entries            /entries               entry      index
+entry              /entries/{id}          entry      show"""
+
         for expected_line, actual_line in zip(expected.splitlines(), str(m).splitlines()):
             assert expected_line == actual_line.rstrip()